requirements.txt
groq==0.10.0
requests==2.32.3
streamlit==1.38.0
python-dotenv==1.0.1